FSK stable send data at 2400 and 4800 speed

+++ Commands/Transmitter/datarate.cpp
@@ -13,7 +13,7 @@ DataRate::DataRate()
 
 void DataRate::setDataRate(const int32_t dataRate)
 {
-   int8_t value =  round(10000000 / 29 / dataRate - 1);
+   int8_t value =  round((double)10000000 / 29 / dataRate) - 1;
    setValue(0, 8, value);
 }
